fix(config): tighten matrix width/height to uint8
Mirrors the Rust u8 pair on MatrixStructureParams (matrix.rs). Rejects out-of-range width/height at Go YAML unmarshal instead of truncating at the Rust boundary. Adds boundary tests (255 ok; 256/300/-1 rejected).

+++ internal/config/types.go
@@ -465,8 +465,8 @@ type StreamConfig struct {
 
 // MatrixStructureParams defines matrix tree shape parameters.
 type MatrixStructureParams struct {
-	Width              int    `yaml:"width" json:"width"`
-	Height             int    `yaml:"height" json:"height"`
+	Width              uint8  `yaml:"width" json:"width"`
+	Height             uint8  `yaml:"height" json:"height"`
 	SpilloverDirection string `yaml:"spillover_direction" json:"spillover_direction"`
 }
